refactor: refactor the ninja controller for include the responseEntity at routes and costomize the messages for customers

+++ src/main/java/com/java10x/cadastrodeninjas/Ninja/Controller/NinjaController.java
@@ -2,6 +2,8 @@ package com.java10x.cadastrodeninjas.Ninja.Controller;
 
 import com.java10x.cadastrodeninjas.Ninja.DTO.NinjaDTO;
 import com.java10x.cadastrodeninjas.Ninja.Service.NinjaService;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -23,33 +25,59 @@ public class NinjaController {
 
     // Add ninja (CREATE)
     @PostMapping("/add")
-    public NinjaDTO addNinja(@RequestBody NinjaDTO ninja) {
-        return ninjaService.addNinja(ninja);
+    public ResponseEntity<String> addNinja(@RequestBody NinjaDTO ninja) {
+        NinjaDTO newNinja = ninjaService.addNinja(ninja);
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body("Ninja created: " + newNinja.getName() + " (ID): " + newNinja.getId());
+
     }
 
     // Get ninjas (READ)
     @GetMapping("/list")
-    public List<NinjaDTO> getNinjas() {
-        return ninjaService.getNinjas();
+    public ResponseEntity<List<NinjaDTO>> getNinjas() {
+        List<NinjaDTO> ninjas = ninjaService.getNinjas();
+        return ResponseEntity.ok(ninjas);
     }
 
     // Get ninja by id  (READ)
     @GetMapping("/list/{id}")
-    public NinjaDTO getNinjaById(@PathVariable Long id) {
-        return ninjaService.getNinjaById(id);
+    public ResponseEntity<?> getNinjaById(@PathVariable Long id) {
+
+        NinjaDTO ninjaById = ninjaService.getNinjaById(id);
+        if (ninjaById != null) {
+            return ResponseEntity.ok(ninjaById);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja not founded!");
+        }
     }
 
     // Modify ninja data (UPDATE)
     @PutMapping("/modify/{id}")
-    public NinjaDTO modifyNinjaById(@PathVariable Long id, @RequestBody NinjaDTO ninjaUpdated) {
-        return ninjaService.modifyNinjaById(id, ninjaUpdated);
+    public ResponseEntity<?> modifyNinjaById(@PathVariable Long id, @RequestBody NinjaDTO ninjaUpdated) {
+
+        NinjaDTO ninjaModify = ninjaService.modifyNinjaById(id, ninjaUpdated);
+
+        if (ninjaModify != null) {
+            return ResponseEntity.ok(ninjaModify);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja not founded!");
+        }
+
     }
 
     // Delete ninja (DELETE)
     @DeleteMapping("/delete/{id}")
-    public String deleteNinjaById(@PathVariable Long id) {
-        ninjaService.deleteNinjaById(id);
-        return "The ninja was deleted!";
+    public ResponseEntity<?> deleteNinjaById(@PathVariable Long id) {
+
+        if (ninjaService.getNinjaById(id) != null) {
+            ninjaService.deleteNinjaById(id);
+            return ResponseEntity.ok("Ninja Deleted, ID: " + id);
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND)
+                    .body("Ninja not founded!");
+        }
     }
 
 }
